requirements.txt
flask==3.0.3
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
pandas==2.2.2
gunicorn==21.2.0

scikit-learn==1.4.2
scipy==1.12.0
tensorflow==2.16.1
keras==3.1.1

python-dateutil==2.9.0.post0
pytz==2024.1

pycoingecko==3.1.0
ccxt==4.2.77

pytest==8.1.1
black==24.3.0
flake8==7.0.0
mypy==1.9.0
pytest-cov==4.1.0
